refactor: (2단계) 테마 및 타이포그래피 적용
- 불필요한 import 제거
- SignUpTitle 을 SignUpLayouts.kt 로 이동
- Surface 의 padding 속성을 SignUpLayout 에 적용
